# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AIOps Platform API
  version: 1.0.0
  description: >-
    Versioned HTTP contract for the AIOps platform. Sensitive credentials,
    passwords, tokens and Kubernetes evidence payloads are intentionally not
    represented in response schemas.
servers:
  - url: /
...
paths:
...
  /api/v1/users:
    get: { tags: [users], summary: List users, operationId: listUsers, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Ok' } } }
    post: { tags: [users], summary: Create user, operationId: createUser, security: [{ bearerAuth: [] }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UserWrite' } } } }, responses: { '201': { $ref: '#/components/responses/User' }, '409': { $ref: '#/components/responses/Error' } } }
  /api/v1/users/assignable:
    get: { tags: [users], summary: List assignable operators, operationId: assignableUsers, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Ok' } } }
  /api/v1/users/{user_id}:
    patch: { tags: [users], summary: Update user status, display name or roles, operationId: updateUser, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/UserID' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UserPatch' } } } }, responses: { '200': { $ref: '#/components/responses/User' }, '409': { $ref: '#/components/responses/Error' } } }
  /api/v1/users/{user_id}/password-reset:
    post: { tags: [users], summary: Reset another user's password, operationId: resetPassword, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/UserID' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PasswordResetRequest' } } } }, responses: { '204': { description: Reset and sessions revoked }, '409': { $ref: '#/components/responses/Error' } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    refreshCookie: { type: apiKey, in: cookie, name: aiops_refresh }
  parameters:
...
    UserID: { name: user_id, in: path, required: true, schema: { type: integer, format: int64 } }
...
  responses:
    Ok: { description: Successful JSON response, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
    User: { description: User response without password material, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
...
    Error: { description: Structured error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  schemas:
...
    Error: { type: object, required: [code, message, request_id], properties: { code: { type: string }, message: { type: string }, request_id: { type: string } } }
...
    User: { type: object, properties: { id: { type: integer }, username: { type: string }, display_name: { type: string }, roles: { type: array, items: { type: string } }, status: { type: string } } }
    UserWrite: { type: object, required: [username, display_name, password, roles], properties: { username: { type: string }, display_name: { type: string }, password: { type: string, format: password }, roles: { type: array, items: { type: string } } } }
    UserPatch: { type: object, properties: { display_name: { type: string }, status: { type: string, enum: [active, disabled] }, roles: { type: array, items: { type: string } } } }
...
    PasswordResetRequest: { type: object, required: [password], properties: { password: { type: string, format: password } } }
...